// packages/webgl/webgl-strict-types/webgl.d.ts
type DOMString = string;
// ...
declare namespace WebGLRenderingContextStrict {
// ...
  interface Extra {
    // Khronos ratified WebGL Extensions
    getExtension(name: 'ANGLE_instanced_arrays '): ANGLE_instanced_arrays | null;
    getExtension(name: 'EXT_blend_minmax '): EXT_blend_minmax | null;
    getExtension(name: 'EXT_frag_depth'): {} | null;
    getExtension(name: 'EXT_shader_texture_lod'): {} | null;
    getExtension(name: 'OES_element_index_uint'): {} | null;
    getExtension(name: 'OES_standard_derivatives'): OES_standard_derivatives | null;
    getExtension(name: 'OES_texture_float'): OES_texture_float | null;
    getExtension(name: 'OES_texture_half_float'): OES_texture_half_float | null;
    getExtension(name: 'OES_texture_half_float_linear'): {} | null;
    getExtension(name: 'OES_vertex_array_object'): OES_vertex_array_object | null;
    getExtension(name: 'WEBGL_depth_texture'): WEBGL_depth_texture | null;
    getExtension(name: 'WEBGL_draw_buffers'): WEBGL_draw_buffers | null;

    // Community approved WebGL Extensions
    getExtension(name: 'EXT_color_buffer_half_float'): EXT_color_buffer_half_float | null;
    getExtension(name: 'EXT_disjoint_timer_query'): EXT_disjoint_timer_query | null;
    getExtension(name: 'EXT_sRGB'): EXT_sRGB | null;
    getExtension(name: 'WEBGL_color_buffer_float'): WEBGL_color_buffer_float | null;

    getActiveAttrib(program: WebGLProgram, index: GLuint): WebGLActiveInfo<AttribType> | null;
    getActiveUniform(program: WebGLProgram, index: GLuint): WebGLActiveInfo<UniformType> | null;

    getRenderbufferParameter(
      target: GL['RENDERBUFFER'],
      pname: GL['RENDERBUFFER_INTERNAL_FORMAT']
    ): RenderbufferInternalFormat;

    getParameter(pname: GL['IMPLEMENTATION_COLOR_READ_FORMAT']): ReadPixelsFormat;
    getParameter(pname: GL['IMPLEMENTATION_COLOR_READ_TYPE']): ReadPixelsType;

    getBufferParameter(target: BufferTarget, pname: GL['BUFFER_USAGE']): BufferDataUsage;

    getUniform(program: WebGLProgram, location: WebGLUniformLocation): UniformType;
  }
// ...
  type ShaderType = GL['FRAGMENT_SHADER'] | GL['VERTEX_SHADER'];
// ...
  type ArrayType = GL['BYTE'] | GL['UNSIGNED_BYTE'] | GL['SHORT'] | GL['UNSIGNED_SHORT'] | GL['FLOAT'];
// ...
  interface Base_ANGLE_instanced_arrays {
    getParameter(pname: WEBGL_draw_buffers['MAX_COLOR_ATTACHMENTS_WEBGL']): GLuint;
    getParameter(pname: WEBGL_draw_buffers['MAX_DRAW_BUFFERS_WEBGL']): GLuint;
    getParameter(pname: DrawBuffer): GL['NONE'] | GL['BACK'] | ColorAttachment; // accurate?
    framebufferRenderbuffer(
      target: GL['FRAMEBUFFER'],
      attachment: ColorAttachment,
      renderbuffertarget: GL['RENDERBUFFER'],
      renderbuffer: WebGLRenderbuffer | null
    ): void;
    framebufferTexture2D(
      target: GL['FRAMEBUFFER'],
      attachment: ColorAttachment,
      textarget: TexImage2DTarget,
      texture: WebGLTexture | null,
      level: GLint
    ): void;
    drawBuffersWEBGL(buffers: Array<GL['NONE'] | GL['BACK'] | ColorAttachment>): void;
    getVertexAttrib(index: GLuint, pname: WEBGL_draw_buffers['MAX_COLOR_ATTACHMENTS_WEBGL']): GLint;
  }
// ...
  interface OES_standard_derivatives {
    readonly FRAGMENT_SHADER_DERIVATIVE_HINT_OES: /* 0x8B8B */ GLenum<'FRAGMENT_SHADER_DERIVATIVE_HINT_OES'>;
  }
  interface Base_OES_standard_derivatives {
    hint(target: OES_standard_derivatives['FRAGMENT_SHADER_DERIVATIVE_HINT_OES'], mode: HintMode): void;
    getParameter(pname: OES_standard_derivatives['FRAGMENT_SHADER_DERIVATIVE_HINT_OES']): HintMode;
  }
// ...
  interface Base_WEBGL_debug_renderer_info {
    getParameter(pname: WEBGL_debug_renderer_info['UNMASKED_VENDOR_WEBGL']): DOMString;
    getParameter(pname: WEBGL_debug_renderer_info['UNMASKED_RENDERER_WEBGL']): DOMString;
  }
// ...
  interface Base_WEBGL_compressed_texture_s3tc {
    /**
     * [MDN Reference](https://developer.mozilla.org/docs/Web/API/WebGLRenderingContext/compressedTexSubImage2D)
     */
    compressedTexImage2D(
      target: TexImage2DTarget,
      level: GLint,
      internalformat: CompressedTextureFormatS3tc,
      width: GLsizei,
      height: GLsizei,
      border: 0,
      /* [AllowShared] */ data: ArrayBufferView
    ): void;
    compressedTexSubImage2D(
      target: TexImage2DTarget,
      level: GLint,
      xoffset: GLint,
      yoffset: GLint,
      width: GLsizei,
      height: GLsizei,
      format: CompressedTextureFormatS3tc,
      /* [AllowShared] */ data: ArrayBufferView
    ): void;
  }
// ...
  interface Base_WEBGL_depth_texture {
    texImage2D(
      target: TexImage2DTarget,
      level: GLint,
      internalformat: GL['DEPTH_COMPONENT'],
      width: GLsizei,
      height: GLsizei,
      border: 0,
      format: GL['DEPTH_COMPONENT'],
      type: GL['UNSIGNED_SHORT'],
      /* [AllowShared] */ pixels: Uint16Array | null
    ): void;
    texImage2D(
      target: TexImage2DTarget,
      level: GLint,
      internalformat: GL['DEPTH_COMPONENT'],
      width: GLsizei,
      height: GLsizei,
      border: 0,
      format: GL['DEPTH_COMPONENT'],
      type: GL['UNSIGNED_INT'],
      /* [AllowShared] */ pixels: Uint32Array | null
    ): void;
    texImage2D(
      target: TexImage2DTarget,
      level: GLint,
      internalformat: GL['DEPTH_STENCIL'],
      width: GLsizei,
      height: GLsizei,
      border: 0,
      format: GL['DEPTH_STENCIL'],
      type: WEBGL_depth_texture['UNSIGNED_INT_24_8_WEBGL'],
      /* [AllowShared] */ pixels: Uint32Array | null
    ): void;

    // May throw DOMException:
    texImage2D(
      target: TexImage2DTarget,
      level: GLint,
      internalformat: GL['DEPTH_COMPONENT'],
      format: GL['DEPTH_COMPONENT'],
      type: GL['UNSIGNED_SHORT'] | GL['UNSIGNED_INT'],
      source: TexImageSource
    ): void;
    texImage2D(
      target: TexImage2DTarget,
      level: GLint,
      internalformat: GL['DEPTH_STENCIL'],
      format: GL['DEPTH_STENCIL'],
      type: WEBGL_depth_texture['UNSIGNED_INT_24_8_WEBGL'],
      source: TexImageSource
    ): void;

    framebufferTexture2D(
      target: GL['FRAMEBUFFER'],
      attachment: GL['DEPTH_STENCIL_ATTACHMENT'],
      textarget: TexImage2DTarget,
      texture: WebGLTexture | null,
      level: GLint
    ): void;
  }

  // https://www.khronos.org/registry/webgl/extensions/Base_OES_element_index_uint/
  interface Base_OES_element_index_uint {
    drawElements(mode: DrawMode, count: GLsizei, type: GL['UNSIGNED_INT'], offset: GLintptr): void;
  }

  // https://www.khronos.org/registry/webgl/extensions/EXT_texture_filter_anisotropic/
  /* [NoInterfaceObject] */
  interface EXT_texture_filter_anisotropic {
    readonly TEXTURE_MAX_ANISOTROPY_EXT: /* 0x84FE */ GLenum<'TEXTURE_MAX_ANISOTROPY_EXT'>;
    readonly MAX_TEXTURE_MAX_ANISOTROPY_EXT: /* 0x84FF */ GLenum<'MAX_TEXTURE_MAX_ANISOTROPY_EXT'>;
  }
  interface Base_EXT_texture_filter_anisotropic {
    getParameter(pname: EXT_texture_filter_anisotropic['MAX_TEXTURE_MAX_ANISOTROPY_EXT']): GLfloat;
    getTexParameter(
      target: TextureTarget,
      pname: EXT_texture_filter_anisotropic['TEXTURE_MAX_ANISOTROPY_EXT']
    ): GLfloat;
    texParameterf(
      target: TextureTarget,
      pname: EXT_texture_filter_anisotropic['TEXTURE_MAX_ANISOTROPY_EXT'],
      param: GLfloat
    ): void;
  }
// ...
  interface Base_WEBGL_draw_buffers {
    getParameter(pname: WEBGL_draw_buffers['MAX_COLOR_ATTACHMENTS_WEBGL']): GLuint;
    getParameter(pname: WEBGL_draw_buffers['MAX_DRAW_BUFFERS_WEBGL']): GLuint;
    getParameter(pname: DrawBuffer): GL['NONE'] | GL['BACK'] | ColorAttachment;
    framebufferRenderbuffer(
      target: GL['FRAMEBUFFER'],
      attachment: ColorAttachment,
      renderbuffertarget: GL['RENDERBUFFER'],
      renderbuffer: WebGLRenderbuffer | null
    ): void;
    framebufferTexture2D(
      target: GL['FRAMEBUFFER'],
      attachment: ColorAttachment,
      textarget: TexImage2DTarget,
      texture: WebGLTexture | null,
      level: GLint
    ): void;
  }
// ...
  interface EXT_blend_minmax {
    readonly MIN_EXT: /* 0x8007 */ GLenum<'MIN_EXT'>;
    readonly MAX_EXT: /* 0x8008 */ GLenum<'MAX_EXT'>;
  }
  interface Base_EXT_blend_minmax {
    blendEquation(mode: BlendEquationModeMinMax): void;
    blendEquationSeparate(modeRGB: BlendEquationModeMinMax, modeAlpha: BlendEquationModeMinMax): void;
  }
// ...
  interface Base_EXT_color_buffer_half_float {
    renderbufferStorage(
      target: GL['RENDERBUFFER'],
      internalformat: EXT_color_buffer_half_float['RGBA16F_EXT'] | EXT_color_buffer_half_float['RGB16F_EXT'],
      width: GLsizei,
      height: GLsizei
    ): void;
  }
// ...
  interface Base_EXT_disjoint_timer_query {
    getParameter(pname: EXT_disjoint_timer_query['TIMESTAMP_EXT']): GLuint64EXT;
    getParameter(pname: EXT_disjoint_timer_query['GPU_DISJOINT_EXT']): boolean;
  }
// ...
  interface Base_EXT_sRGB {
    texImage2D(
      target: TexImage2DTarget,
      level: GLint,
      internalformat: EXT_sRGB['SRGB_EXT'],
      width: GLsizei,
      height: GLsizei,
      border: 0,
      format: EXT_sRGB['SRGB_EXT'],
      type: GL['UNSIGNED_BYTE'],
      /* [AllowShared] */ pixels: Uint8Array | null
    ): void;
    texImage2D(
      target: TexImage2DTarget,
      level: GLint,
      internalformat: EXT_sRGB['SRGB_ALPHA_EXT'],
      width: GLsizei,
      height: GLsizei,
      border: 0,
      format: EXT_sRGB['SRGB_ALPHA_EXT'],
      type: GL['UNSIGNED_BYTE'],
      /* [AllowShared] */ pixels: Uint8Array | null
    ): void;

    texImage2D(
      target: TexImage2DTarget,
      level: GLint,
      internalformat: EXT_sRGB['SRGB_EXT'],
      format: EXT_sRGB['SRGB_EXT'],
      type: GL['UNSIGNED_BYTE'],
      source: TexImageSource
    ): void;
    texImage2D(
      target: TexImage2DTarget,
      level: GLint,
      internalformat: EXT_sRGB['SRGB_ALPHA_EXT'],
      format: EXT_sRGB['SRGB_ALPHA_EXT'],
      type: GL['UNSIGNED_BYTE'],
      source: TexImageSource
    ): void;

    texSubImage2D(
      target: TexImage2DTarget,
      level: GLint,
      xoffset: GLint,
      yoffset: GLint,
      width: GLsizei,
      height: GLsizei,
      format: EXT_sRGB['SRGB_EXT'] | EXT_sRGB['SRGB_ALPHA_EXT'],
      type: GL['UNSIGNED_BYTE'],
      /* [AllowShared] */ pixels: Uint8Array | null
    ): void;
    texSubImage2D(
      target: TexImage2DTarget,
      level: GLint,
      xoffset: GLint,
      yoffset: GLint,
      format: EXT_sRGB['SRGB_EXT'] | EXT_sRGB['SRGB_ALPHA_EXT'],
      type: GL['UNSIGNED_BYTE'],
      source: TexImageSource
    ): void;

    renderbufferStorage(
      target: GL['RENDERBUFFER'],
      internalformat: EXT_sRGB['SRGB8_ALPHA8_EXT'],
      width: GLsizei,
      height: GLsizei
    ): void;

    getFramebufferAttachmentParameter(
      target: GL['FRAMEBUFFER'],
      attachment: ColorAttachment,
      pname: EXT_sRGB['FRAMEBUFFER_ATTACHMENT_COLOR_ENCODING_EXT']
    ): GL['LINEAR'] | EXT_sRGB['SRGB_EXT'];
  }
// ...
  interface Base_WEBGL_color_buffer_float {
    renderbufferStorage(
      target: GL['RENDERBUFFER'],
      internalformat: WEBGL_color_buffer_float['RGBA32F_EXT'] | WEBGL_color_buffer_float['RGB32F_EXT'],
      width: GLsizei,
      height: GLsizei
    ): void;
  }
// ...
}

type WebGLRenderingContextStrict = WebGLRenderingContextStrict.Base &
  WebGLRenderingContextStrict.Extra &
  WebGLRenderingContextStrict.Base_ANGLE_instanced_arrays &
  WebGLRenderingContextStrict.Base_EXT_blend_minmax &
  WebGLRenderingContextStrict.Base_EXT_texture_filter_anisotropic &
  WebGLRenderingContextStrict.Base_OES_element_index_uint &
  WebGLRenderingContextStrict.Base_OES_standard_derivatives &
  WebGLRenderingContextStrict.Base_OES_texture_float &
  WebGLRenderingContextStrict.Base_OES_texture_half_float &
  WebGLRenderingContextStrict.Base_WEBGL_compressed_texture_s3tc &
  WebGLRenderingContextStrict.Base_WEBGL_debug_renderer_info &
  WebGLRenderingContextStrict.Base_WEBGL_depth_texture &
  WebGLRenderingContextStrict.Base_WEBGL_draw_buffers &
  WebGLRenderingContextStrict.Base_EXT_color_buffer_half_float &
  WebGLRenderingContextStrict.Base_EXT_disjoint_timer_query &
  WebGLRenderingContextStrict.Base_EXT_sRGB &
  WebGLRenderingContextStrict.Base_WEBGL_color_buffer_float &
  WebGLRenderingContextStrict.Base_WEBGL_compressed_texture_astc &
  WebGLRenderingContextStrict.Base_WEBGL_compressed_texture_atc &
  WebGLRenderingContextStrict.Base_WEBGL_compressed_texture_etc &
  WebGLRenderingContextStrict.Base_WEBGL_compressed_texture_etc1 &
  WebGLRenderingContextStrict.Base_WEBGL_compressed_texture_pvrtc &
  WebGLRenderingContextStrict.Base_WEBGL_compressed_texture_s3tc_srgb;